cambios en la bbdd

+++ back/kalio-api/database/migrations/2025_03_05_155642_create_consum_diaris_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('repte_id')->constrained('reptes')->onDelete('cascade'); // Relación con el reto
             $table->date('data'); // Fecha del consumo
-            $table->integer('calories_consumides'); // Cantidad de calorías consumidas
-            $table->integer('proteines_consumides')->nullable(); // Cantidad de proteínas consumidas
-            $table->integer('sucres_consumits')->nullable(); // Cantidad de azúcares consumidos
+            $table->float('calories_consumides'); // Cantidad de calorías consumidas
+            $table->float('proteines_consumides')->nullable(); // Cantidad de proteínas consumidas
+            $table->float('sucres_consumits')->nullable(); // Cantidad de azúcares consumidos
             $table->timestamps();
         });
     }
